Refactor Form Construction
